infinite addition

+++ python-import_modules/3-infinite_add.py
@@ -5,5 +5,5 @@ if __name__ == "__main__":
     add = 0
     n = len(sys.argv)
     for i in range (1, n):
-        add += float(sys.argv[i])
+        add += int(sys.argv[i])
     print(add)
